[refactor] Use SGLangEngine directly and add profiler endpoint (#139)
