+++ slime/backends/sglang_utils/sglang_engine.py
@@ -1,14 +1,19 @@
 import dataclasses
 
 import os
-from typing import TYPE_CHECKING
 
 from sglang.srt.server_args import ServerArgs
 from slime.utils.http_utils import get_host_info
-from .http_server_engine import HttpServerEngineAdapter
+import multiprocessing
+import time
+from typing import List, Optional
 
-if TYPE_CHECKING:
-    pass
+import requests
+from sglang.srt.entrypoints.http_server import launch_server
+from sglang.srt.server_args import ServerArgs
+from sglang.srt.utils import kill_process_tree
+from urllib3.exceptions import NewConnectionError
+from slime.ray.ray_actor import RayActor
 
 
 def get_base_gpu_id(args, rank):
@@ -21,10 +26,61 @@ def get_base_gpu_id(args, rank):
     return start_index
 
 
-class SglangEngine:
+def launch_server_process(server_args: ServerArgs) -> multiprocessing.Process:
 
-    def __init__(self, args, rank, dist_init_addr, port, nccl_port):
+    p = multiprocessing.Process(target=launch_server, args=(server_args,))
+    p.start()
+
+    if server_args.node_rank != 0:
+        return
+
+    base_url = server_args.url()
+
+    headers = {
+        "Content-Type": "application/json; charset=utf-8",
+        "Authorization": f"Bearer {server_args.api_key}",
+    }
+
+    with requests.Session() as session:
+        while True:
+            try:
+                response = session.get(f"{base_url}/health_generate", headers=headers)
+                if response.status_code == 200:
+                    break
+            except requests.RequestException:
+                pass
+
+            if not p.is_alive():
+                raise Exception("Server process terminated unexpectedly.")
+
+            time.sleep(2)
+
+        # use flush_cache to make sure the working queue is empty, so that we can do offload
+        while True:
+            try:
+                response = session.get(f"{base_url}/flush_cache", headers=headers)
+                if response.status_code == 200:
+                    break
+
+            except requests.RequestException:
+                pass
+
+            if not p.is_alive():
+                raise Exception("Server process terminated unexpectedly.")
+
+            time.sleep(2)
+
+    return p
+
+
+class SGLangEngine(RayActor):
+    def __init__(self, args, rank: int):
         self.args = args
+        self.rank = rank
+
+    def init(self, dist_init_addr, port, nccl_port):
+        args = self.args
+        rank = self.rank
 
         # remove the CUDA_VISIBLE_DEVICES set by ray and use base_gpu_id
         os.environ.pop("CUDA_VISIBLE_DEVICES", None)
@@ -67,36 +123,146 @@ class SglangEngine:
             for key in unused_keys:
                 kwargs.pop(key)
 
-        self.llm = HttpServerEngineAdapter(
-            router_ip=args.sglang_router_ip, router_port=args.sglang_router_port, **kwargs
+        self.router_ip = args.sglang_router_ip
+        self.router_port = args.sglang_router_port
+        self.server_args = ServerArgs(**kwargs)
+        self.node_rank = self.server_args.node_rank
+        print(f"Launch HttpServerEngineAdapter at: {self.server_args.host}:{self.server_args.port}")
+        self.process = launch_server_process(self.server_args)
+        if self.node_rank == 0 and self.router_ip and self.router_port:
+            requests.post(
+                f"http://{self.router_ip}:{self.router_port}/add_worker?url=http://{self.server_args.host}:{self.server_args.port}"
+            )
+
+        if self.args.offload:
+            # offload the engine to the CPU
+            self.release_memory_occupation()
+
+    def _make_request(self, endpoint: str, payload: Optional[dict] = None):
+        """Make a POST request to the specified endpoint with the given payload.
+
+        Args:
+            endpoint: The API endpoint to call
+            payload: The JSON payload to send (default: empty dict)
+
+        Returns:
+            The JSON response from the server
+        """
+        if self.node_rank != 0:
+            return
+
+        url = f"http://{self.server_args.host}:{self.server_args.port}/{endpoint}"
+        response = requests.post(url, json=payload or {})
+        response.raise_for_status()
+        return response.json()
+
+    def update_weights_from_tensor(
+        self,
+        serialized_named_tensors: List[str],
+        load_format: Optional[str] = None,
+        flush_cache: bool = False,
+    ):
+        """
+        Update model weights from tensor data. The HTTP server will only post meta data, and the real weights will be copied directly from GPUs.
+
+        Note: The model should be on GPUs rather than CPU for this functionality to work properly.
+        If you encounter issues, ensure your model is loaded on GPU devices rather than CPU.
+        """
+
+        return self._make_request(
+            "update_weights_from_tensor",
+            {
+                "serialized_named_tensors": serialized_named_tensors,
+                "load_format": load_format,
+                "flush_cache": flush_cache,
+            },
         )
 
-    def init_process_group(self, master_address, master_port, rank_offset, world_size, group_name, backend):
-        return self.llm.init_weights_update_group(
-            master_address, master_port, rank_offset, world_size, group_name, backend
+    def flush_cache(self):
+        """Flush the cache of the server."""
+        if self.node_rank != 0:
+            return
+        # flush cache will not return status_code 200 when there are pending requests
+        while True:
+            try:
+                response = requests.get(f"http://{self.server_args.host}:{self.server_args.port}/flush_cache")
+                if response.status_code == 200:
+                    break
+            except NewConnectionError as e:
+                raise e
+            except Exception as e:
+                print(f"Error flushing cache: {e}")
+                continue
+
+    def shutdown(self):
+        requests.post(
+            f"http://{self.router_ip}:{self.router_port}/remove_worker?url=http://{self.server_args.host}:{self.server_args.port}"
+        )
+        kill_process_tree(self.process.pid)
+
+    def release_memory_occupation(self):
+        return self._make_request("release_memory_occupation")
+
+    def resume_memory_occupation(self):
+        return self._make_request("resume_memory_occupation")
+
+    def init_weights_update_group(self, master_address, master_port, rank_offset, world_size, group_name, backend):
+        return self._make_request(
+            "init_weights_update_group",
+            {
+                "master_address": master_address,
+                "master_port": master_port,
+                "rank_offset": rank_offset,
+                "world_size": world_size,
+                "group_name": group_name,
+                "backend": backend,
+            },
         )
 
-    def update_weights_from_distributed(self, names, dtypes, shapes, group_name):
-        self.llm.update_weights_from_distributed(names, dtypes, shapes, group_name)
-        return
-
-    def update_weights_from_tensor(self, ipc_handles):
-        self.llm.update_weights_from_tensor(ipc_handles)
-        return
-
-    def reset_prefix_cache(self):
-        self.llm.flush_cache()
-
-    def sleep(self, level=1):
-        # Adhoc solution to ensure no running requests
-        self.llm.flush_cache()
-        self.llm.release_memory_occupation()
-
-    def wake_up(self):
-        self.llm.resume_memory_occupation()
+    def update_weights_from_distributed(self, names, dtypes, shapes, group_name, flush_cache=False):
+        return self._make_request(
+            "update_weights_from_distributed",
+            {
+                "names": names,
+                "dtypes": [str(dtype).replace("torch.", "") for dtype in dtypes],
+                "shapes": shapes,
+                "group_name": group_name,
+                "flush_cache": flush_cache,
+            },
+        )
 
     def pause_generation(self):
-        self.llm.pause_generation()
+        return requests.post(f"http://{self.server_args.host}:{self.server_args.port}/pause_generation", json={})
 
     def continue_generation(self):
-        self.llm.continue_generation()
+        return requests.post(f"http://{self.server_args.host}:{self.server_args.port}/continue_generation", json={})
+
+    def start_profile(
+        self,
+        # The output directory
+        output_dir: Optional[str] = None,
+        # If set, it profile as many as this number of steps.
+        # If it is set, profiling is automatically stopped after this step, and
+        # the caller doesn't need to run stop_profile.
+        start_step: Optional[int] = None,
+        num_steps: Optional[int] = None,
+        activities: Optional[List[str]] = None,
+        profile_by_stage: bool = False,
+        with_stack: Optional[bool] = None,
+        record_shapes: Optional[bool] = None,
+    ):
+        return requests.post(
+            f"http://{self.server_args.host}:{self.server_args.port}/start_profile",
+            json={
+                "output_dir": output_dir,
+                "start_step": start_step,
+                "num_steps": num_steps,
+                "activities": activities,
+                "profile_by_stage": profile_by_stage,
+                "with_stack": with_stack,
+                "record_shapes": record_shapes,
+            },
+        )
+
+    def stop_profile(self):
+        return requests.post(f"http://{self.server_args.host}:{self.server_args.port}/stop_profile", json={})
