+++ slime/ray/rollout.py
@@ -5,58 +5,10 @@ import time
 import ray
 from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
 
-from slime.backends.sglang_utils.sglang_engine import SglangEngine
+from slime.backends.sglang_utils.sglang_engine import SGLangEngine
 from slime.ray.buffer import Buffer
-from slime.ray.ray_actor import RayActor
 from slime.utils.http_utils import find_available_port, get_host_info, run_router
 from .utils import Lock
-
-
-@ray.remote
-class RolloutRayActor(RayActor):
-    def __init__(self, args, rank: int):
-        self.args = args
-        self.rank = rank
-
-    def init(self, dist_init_addr, port, nccl_port):
-        # build infer engine
-        self.infer_engine = SglangEngine(
-            args=self.args,
-            rank=self.rank,
-            dist_init_addr=dist_init_addr,
-            port=port,
-            nccl_port=nccl_port,
-        )
-
-        if self.args.offload:
-            # offload the engine to the CPU
-            self.infer_engine.sleep()
-
-    def init_process_group(self, master_address, master_port, rank_offset, world_size, group_name, backend):
-        return self.infer_engine.init_process_group(
-            master_address, master_port, rank_offset, world_size, group_name, backend
-        )
-
-    def update_weights_from_distributed(self, names, dtypes, shapes, group_name):
-        return self.infer_engine.update_weights_from_distributed(names, dtypes, shapes, group_name)
-
-    def update_weights_from_tensor(self, ipc_handles):
-        return self.infer_engine.update_weights_from_tensor(ipc_handles)
-
-    def reset_prefix_cache(self):
-        self.infer_engine.reset_prefix_cache()
-
-    def sleep(self, level=1):
-        self.infer_engine.sleep(level=level)
-
-    def wake_up(self):
-        self.infer_engine.wake_up()
-
-    def pause_generation(self):
-        self.infer_engine.pause_generation()
-
-    def continue_generation(self):
-        self.infer_engine.continue_generation()
 
 
 def create_rollout_engines(args, pg):
@@ -67,6 +19,8 @@ def create_rollout_engines(args, pg):
     num_engines = args.rollout_num_gpus // num_gpu_per_engine
 
     pg, reordered_bundle_indices = pg
+
+    RolloutRayActor = ray.remote(SGLangEngine)
 
     rollout_engines = []
     for i in range(num_engines):
@@ -207,11 +161,8 @@ class RolloutManager:
     def async_generate(self, rollout_id, evaluation=False):
         return self.data_buffer.generate.remote(rollout_id, evaluation=evaluation)
 
-    def async_reset_prefix_cache(self):
-        return [engine.reset_prefix_cache.remote() for engine in self.rollout_engines]
-
     def async_offload(self):
-        return [engine.sleep.remote() for engine in self.rollout_engines]
+        return [engine.release_memory_occupation.remote() for engine in self.rollout_engines]
 
     def async_onload(self):
-        return [engine.wake_up.remote() for engine in self.rollout_engines]
+        return [engine.resume_memory_occupation.remote() for engine in self.rollout_engines]
